Respond to review comment.
This also does a few extra things:

 - Depend on cats-core
 - Impelment contravariant and (non-default) cartesian
 - Split Document to its own file
